Better handling for enum values; EnumEditorScreen
