+++ engine/src/main/java/org/terasology/rendering/nui/editor/EnumEditorScreen.java
@@ -17,40 +17,29 @@ package org.terasology.rendering.nui.editor;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 import org.terasology.assets.ResourceUrn;
-import org.terasology.reflection.metadata.ClassLibrary;
-import org.terasology.reflection.metadata.ClassMetadata;
 import org.terasology.rendering.nui.CoreScreenLayer;
-import org.terasology.rendering.nui.UILayout;
-import org.terasology.rendering.nui.UIWidget;
 import org.terasology.rendering.nui.WidgetUtil;
 import org.terasology.rendering.nui.databinding.ReadOnlyBinding;
 import org.terasology.rendering.nui.widgets.UIButton;
 import org.terasology.rendering.nui.widgets.UIDropdownScrollable;
 import org.terasology.rendering.nui.widgets.UpdateListener;
 import org.terasology.rendering.nui.widgets.treeView.JsonTree;
-import org.terasology.rendering.nui.widgets.treeView.JsonTreeValue;
 
-import java.util.Collections;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
-public class WidgetSelectionScreen extends CoreScreenLayer {
-    public static final ResourceUrn ASSET_URI = new ResourceUrn("engine:widgetSelectionScreen");
+public class EnumEditorScreen extends CoreScreenLayer {
+    public static final ResourceUrn ASSET_URI = new ResourceUrn("engine:enumEditorScreen");
 
     /**
-     * The dropdown containing the values of {@code widgets}.
+     * The dropdown containing the possible values of an enum.
      */
-    private UIDropdownScrollable availableWidgets;
+    private UIDropdownScrollable enumValues;
     /**
-     * The {@link JsonTree} a selected widget is to be added to.
+     *
      */
     private JsonTree node;
-    /**
-     * The list of available UIWidget instances, excluding CoreScreenLayer overrides.
-     */
-    private Map<String, ClassMetadata> widgets = Maps.newHashMap();
     /**
      * Listeners fired when the screen is closed (without selecting a widget).
      */
@@ -58,34 +47,10 @@ public class WidgetSelectionScreen extends CoreScreenLayer {
 
     @Override
     public void initialise() {
-        availableWidgets = find("availableWidgets", UIDropdownScrollable.class);
-
-        // Populate the widget list.
-        ClassLibrary<UIWidget> metadataLibrary = getManager().getWidgetMetadataLibrary();
-        for (ClassMetadata metadata : metadataLibrary) {
-            if (!CoreScreenLayer.class.isAssignableFrom(metadata.getType())) {
-                widgets.put(metadata.toString(), metadata);
-            }
-        }
-
-        List options = Lists.newArrayList(widgets.keySet());
-        Collections.sort(options);
-        availableWidgets.setOptions(options);
+        enumValues = find("enumValues", UIDropdownScrollable.class);
 
         WidgetUtil.trySubscribe(this, "ok", button -> {
-            String selection = availableWidgets.getSelection().toString();
-
-            ClassMetadata metadata = widgets.get(selection);
-
-            JsonTree childNode = NUIEditorNodeUtils.createNewWidget(selection, "newWidget", false);
-
-            // If the widget is an UILayout override, add a "contents" array node.
-            if (UILayout.class.isAssignableFrom(metadata.getType())) {
-                childNode.addChild(new JsonTreeValue("contents", null, JsonTreeValue.Type.ARRAY));
-            }
-
-            node.addChild(childNode);
-
+            node.getValue().setValue(enumValues.getSelection().toString());
             closeListeners.forEach(UpdateListener::onAction);
             getManager().closeScreen(ASSET_URI);
         });
@@ -93,7 +58,7 @@ public class WidgetSelectionScreen extends CoreScreenLayer {
         find("ok", UIButton.class).bindEnabled(new ReadOnlyBinding<Boolean>() {
             @Override
             public Boolean get() {
-                return availableWidgets.getSelection() != null;
+                return enumValues.getSelection() != null;
             }
         });
 
@@ -105,6 +70,11 @@ public class WidgetSelectionScreen extends CoreScreenLayer {
     public void setNode(JsonTree node) {
         this.node = node;
     }
+
+    public void setEnumClass(Class clazz) {
+        enumValues.setOptions(Arrays.asList(clazz.getEnumConstants()));
+    }
+
 
     public void subscribeClose(UpdateListener listener) {
         Preconditions.checkNotNull(listener);
